requirements.txt
Click >= 6.6
PyYAML >= 3.11
